requirements.txt
folia-tools
pandas
tqdm
transformers
PyPDF2
python-ucto
matplotlib
os